refactor(typecript): move cript types to project root
Also added a path for cript easy import @cript
