+++ scripts/typescript/src/cript/EditorMode.ts
