+++ scripts/typescript/src/cript/IAlgorithm.ts
