+++ scripts/typescript/src/cript/ICollection..ts
