+++ scripts/typescript/src/cript/IComputation.ts
