+++ scripts/typescript/src/cript/IComputationProcess.ts
