+++ scripts/typescript/src/cript/ICondition.d.ts
