+++ scripts/typescript/src/cript/IData.ts
