+++ scripts/typescript/src/cript/IEquipment.ts
