+++ scripts/typescript/src/cript/IExperiment.ts
