+++ scripts/typescript/src/cript/IFile.ts
