+++ scripts/typescript/src/cript/IIdentifier.ts
