+++ scripts/typescript/src/cript/IIngredient.ts
