+++ scripts/typescript/src/cript/IInventory.ts
