+++ scripts/typescript/src/cript/IMaterial.ts
