+++ scripts/typescript/src/cript/IParameter.ts
