+++ scripts/typescript/src/cript/IProcess.ts
