+++ scripts/typescript/src/cript/IProject.ts
