+++ scripts/typescript/src/cript/IProperty.ts
