+++ scripts/typescript/src/cript/IQuantity.ts
