+++ scripts/typescript/src/cript/IReference.ts
