+++ scripts/typescript/src/cript/ISoftware.ts
