+++ scripts/typescript/src/cript/ISoftwareConfiguration.ts
