+++ scripts/typescript/src/cript/IUser.ts
