+++ scripts/typescript/src/cript/IVocab.ts
