+++ scripts/typescript/src/cript/index.ts
